requirements.txt
fastapi
uvicorn
python-multipart
jinja2
motor
python-dotenv
requests
sendgrid==6.9.7
certifi==2024.2.2